requirements.txt
coreapi==2.3.3
dj-database-url==0.5.0
Django==2.2.6
django-heroku==0.3.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2==2.8.4
python-decouple==3.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
Pillow==6.2.1
